fix(hooks): scan gh -F/--body-file short-form body file on the publish gate (#1812)
gh issue/pr create|comment's -F is the documented short form of
--body-file <file>, but the body extractor only treated -F as a file
reference for git; for gh/glab it routed -F to the api-field walker
(which expects -F name=value), so a bare -F <path> matched no walker and
the file content was never read. A banned overlay/customer term inside a
-F body file posted to a PUBLIC target therefore bypassed the
banned-terms / bare-reference gates unscanned.

Disambiguate the overloaded short -F by value: a =-free token is a
--body-file path (read and scanned, fail-closed-on-unreadable for the
destination-aware gates); a name=value token stays an api field
assignment. git -F is unchanged (always a file). Both the space-separated
-F <path> and attached -F<path> spellings are covered.

Extract the body/payload argument walkers (flag catalogues, body-flag /
body-file / api-field / curl walkers, heredoc resolution, the fail-closed
sentinel) into a dedicated _arg_body_walkers module so _command_parser
stays under the per-file LOC ceiling and the parser facade keeps only
detection, segment routing, and the public extract_* entry points.

Golden must-DENY + must-ALLOW pair through the live
handle_banned_terms_pretool entry point and the destination-aware corpus;
must-DENY proven RED before the fix. 100% coverage on the new lines; full
suite green at 94.41%.
