+++ src/teatree/hooks/_command_parser.py
@@ -31,11 +31,20 @@ classifiers there, so an interspersed persistent flag cannot break detection
 (#1672). This module owns body / title / secret-surface EXTRACTION.
 """
 
-import json
 import re
-from pathlib import Path
 from typing import Final
 
+from teatree.hooks._arg_body_walkers import (
+    FAIL_CLOSED_SENTINEL,
+    api_field_values,
+    heredoc_file_bodies,
+    heredoc_inline_bodies,
+    is_fail_closed_sentinel,
+    walk_api_fields,
+    walk_body_file_flags,
+    walk_body_flags,
+    walk_curl_args,
+)
 from teatree.hooks._publish_detection import (
     command_has_opaque_forge_transport,
     command_has_token_aware_publish_surface,
@@ -43,6 +52,17 @@ from teatree.hooks._publish_detection import (
     segment_word_lists,
 )
 from teatree.hooks._shell_lexer import Token, TokenKind, is_command_separator, split_commands, tokenize
+
+__all__ = [
+    "FAIL_CLOSED_SENTINEL",
+    "extract_bash_payload",
+    "extract_secret_scan_text",
+    "extract_title_fragments",
+    "first_segment_words",
+    "is_fail_closed_sentinel",
+    "is_publish_command",
+    "normalize_for_substring_match",
+]
 
 # ── Publish-surface substring catalogues ────────────────────────────
 
@@ -97,31 +117,6 @@ _T3_PUBLISH_SUBSTRINGS: Final[tuple[str, ...]] = (
 )
 
 
-# Sentinel string that downstream scanning treats as a HIGH match. Any
-# indirect or undecodable body source surfaces this so the gate fails
-# closed (codex CRITICAL #5 round 1, codex round-2 #4).
-#
-# The wording must NOT itself match any quote-scanner HIGH pattern,
-# otherwise the gate self-matches its own injected sentinel and reports
-# a bogus user-quote finding on a body it never actually saw (#126: the
-# old "the user said: …" phrasing tripped ``the-user-said-colon``). The
-# sentinel is recognised explicitly by :func:`is_fail_closed_sentinel`
-# so a scanner can fail closed on a NAMED reason instead of an
-# accidental content-pattern self-match.
-FAIL_CLOSED_SENTINEL: Final[str] = "[teatree-gate] pre-publish scanner could not resolve a body source; failing closed"
-
-
-def is_fail_closed_sentinel(text: str) -> bool:
-    """Return True iff ``text`` carries the injected fail-closed sentinel.
-
-    Callers fail closed on a NAMED reason rather than rely on the
-    sentinel accidentally tripping a content pattern (#126). The sentinel
-    can be one of several concatenated payload fragments, so a substring
-    test is used rather than equality.
-    """
-    return FAIL_CLOSED_SENTINEL in text
-
-
 def normalize_for_substring_match(command: str) -> str:
     r"""Return a publish-detection-friendly string for ``command``.
 
@@ -166,345 +161,6 @@ def is_publish_command(command: str) -> bool:
     return command_has_token_aware_publish_surface(command)
 
 
-# ── Body-flag and curl regexes (heredoc only — flag args are token-aware) ─
-
-_HEREDOC_RE: Final[re.Pattern[str]] = re.compile(
-    r"<<\s*['\"]?(\w+)['\"]?\s*\n(.*?)\n\1\b",
-    re.DOTALL,
-)
-
-# A redirect (``> path`` / ``>| path`` / ``>> path``) that writes a
-# heredoc body to a file, e.g. ``cat > /tmp/msg.txt <<'EOF' … EOF``. The
-# common agent idiom is to write a commit message to a temp file and
-# then ``git commit -F /tmp/msg.txt`` — at PreToolUse scan time that file
-# does NOT exist yet (the hook runs BEFORE the command), so the only
-# place the body lives is the in-command heredoc. This regex pairs the
-# redirect target path with the heredoc delimiter so :func:`extract_bash_payload`
-# can resolve a ``-F <path>`` reference to the body the command is about
-# to write there (#126).
-_HEREDOC_TO_FILE_RE: Final[re.Pattern[str]] = re.compile(
-    r">{1,2}\|?\s*(?P<path>'[^']+'|\"[^\"]+\"|\S+)\s+<<\s*['\"]?(?P<delim>\w+)['\"]?\s*\n(?P<body>.*?)\n(?P=delim)\b",
-    re.DOTALL,
-)
-
-
-def _heredoc_file_bodies(command: str) -> dict[str, str]:
-    """Map each ``> path <<EOF … EOF`` redirect target to its heredoc body.
-
-    Resolves the agent idiom of writing a body to a temp file and then
-    referencing it via ``git commit -F <path>``. The path is normalised
-    (surrounding quotes stripped) so a quoted redirect target matches the
-    later bare ``-F`` reference (#126).
-    """
-    bodies: dict[str, str] = {}
-    for match in _HEREDOC_TO_FILE_RE.finditer(command):
-        raw_path = match.group("path")
-        path = raw_path.strip("'\"")
-        bodies[path] = match.group("body")
-    return bodies
-
-
-# Per-command argument-walker dispatch tables --------------------------
-
-# Body-bearing long options (value follows the flag as next token or
-# attached via ``=``). The catalogue is shared by all publishing
-# commands — gh, glab, git, curl all use the same long-option grammar.
-_BODY_FLAG_NAMES: Final[frozenset[str]] = frozenset(
-    {"--body", "--description", "--message", "--title"},
-)
-
-# Long options that point at a FILE whose content we should read. If the
-# file is missing or unreadable the parser appends the fail-closed
-# sentinel.
-_BODY_FILE_FLAG_NAMES: Final[frozenset[str]] = frozenset(
-    {"--body-file", "--description-file", "--file"},
-)
-
-# Short body-bearing flags used by ``gh`` / ``glab`` / ``git commit``.
-_BODY_SHORT_FLAGS: Final[frozenset[str]] = frozenset({"-m", "-b"})
-
-# Long options for ``gh api`` / ``glab api`` field assignments.
-_API_FIELD_LONG_FLAGS: Final[frozenset[str]] = frozenset({"--field", "--raw-field"})
-_API_FIELD_SHORT_FLAGS: Final[frozenset[str]] = frozenset({"-f", "-F"})
-
-# Curl long-option data flags — payload is JSON-or-text.
-_CURL_DATA_LONG_FLAGS: Final[frozenset[str]] = frozenset(
-    {"--data", "--data-raw", "--data-binary", "--data-urlencode", "--json"},
-)
-
-
-def _read_file_arg(path: str) -> str | None:
-    try:
-        return Path(path).read_text(encoding="utf-8")
-    except (OSError, UnicodeDecodeError):
-        return None
-
-
-def _json_body_fields(blob: str) -> list[str]:
-    """Return ``text``/``message``/``body`` values from a JSON blob, if any."""
-    try:
-        decoded = json.loads(blob)
-    except (ValueError, TypeError):
-        return []
-    if not isinstance(decoded, dict):
-        return []
-    return [str(decoded[key]) for key in ("text", "message", "body") if key in decoded]
-
-
-def _attached_value(token: str, prefix: str) -> str | None:
-    """Return the attached value of ``-X<value>`` / ``-X=<value>``, if any.
-
-    Returns the substring AFTER ``prefix`` when ``token`` starts with the
-    prefix and is strictly longer than it. ``-X=value`` strips the
-    leading ``=`` so callers see the bare payload.
-    """
-    if token.startswith(prefix) and len(token) > len(prefix):
-        return token[len(prefix) :].removeprefix("=")
-    return None
-
-
-def _scan_curl_payload(raw: str, payloads: list[str]) -> None:
-    """Append ``raw`` plus its JSON ``text``/``message``/``body`` fields.
-
-    A non-JSON-decodable body that LOOKS like JSON (starts with ``{`` or
-    ``[``) fails closed because we cannot be sure the gate's pattern
-    catalogue covers the obfuscation.
-    """
-    payloads.append(raw)
-    json_fields = _json_body_fields(raw)
-    if json_fields:
-        payloads.extend(json_fields)
-    elif raw.strip().startswith(("{", "[")):
-        payloads.append(FAIL_CLOSED_SENTINEL)
-
-
-def _record_curl_value(value: str, payloads: list[str]) -> None:
-    """Route a single curl data value to the payload list.
-
-    ``@file`` references fail closed (we cannot read arbitrary files);
-    everything else gets the standard JSON-aware scan.
-    """
-    if value.startswith("@"):
-        payloads.append(FAIL_CLOSED_SENTINEL)
-    else:
-        _scan_curl_payload(value, payloads)
-
-
-def _curl_long_flag_attached(word: str) -> str | None:
-    """Return the value of ``--data=VALUE`` / ``--json=VALUE`` if attached."""
-    for flag in _CURL_DATA_LONG_FLAGS:
-        attached = _attached_value(word, flag + "=")
-        if attached is not None:
-            return attached
-    return None
-
-
-def _curl_short_d_attached(word: str) -> str | None:
-    """Return the value of ``-dVALUE`` attached short option, if applicable.
-
-    Excludes the long ``--data*`` / ``--json`` family — those start with
-    ``--`` and are handled by :func:`_curl_long_flag_attached`.
-    """
-    if word.startswith(("--data", "--json")):
-        return None
-    return _attached_value(word, "-d")
-
-
-def _walk_curl_args(words: list[str], payloads: list[str]) -> None:
-    """Extract curl ``-d``/``--data*``/``--json`` payloads from a command.
-
-    Supports:
-    - ``-d value`` (next token)
-    - ``-dvalue`` (attached short option, POSIX)
-    - ``-d=value`` (equals form)
-    - ``--data value`` / ``--data=value``
-    - ``-d@file`` / ``--data @file`` (fail closed — we cannot read the file)
-    """
-    i = 0
-    n = len(words)
-    while i < n:
-        word = words[i]
-        if word == "-d" and i + 1 < n:
-            _record_curl_value(words[i + 1], payloads)
-            i += 2
-            continue
-        if word in _CURL_DATA_LONG_FLAGS and i + 1 < n:
-            _record_curl_value(words[i + 1], payloads)
-            i += 2
-            continue
-        attached_short = _curl_short_d_attached(word)
-        if attached_short is not None:
-            _record_curl_value(attached_short, payloads)
-            i += 1
-            continue
-        attached_long = _curl_long_flag_attached(word)
-        if attached_long is not None:
-            _record_curl_value(attached_long, payloads)
-        i += 1
-
-
-def _walk_body_flags(words: list[str], payloads: list[str]) -> None:
-    """Extract ``--body``/``-m``/``-b`` style payloads from a command.
-
-    Handles both space-separated (``--body "x"``) and equals-separated
-    (``--body=x``) forms.
-    """
-    i = 0
-    n = len(words)
-    while i < n:
-        word = words[i]
-        if word in _BODY_FLAG_NAMES and i + 1 < n:
-            payloads.append(words[i + 1])
-            i += 2
-            continue
-        for flag in _BODY_FLAG_NAMES:
-            attached = _attached_value(word, flag + "=")
-            if attached is not None:
-                payloads.append(attached)
-                break
-        if word in _BODY_SHORT_FLAGS and i + 1 < n:
-            payloads.append(words[i + 1])
-            i += 2
-            continue
-        i += 1
-
-
-def _walk_body_file_flags(
-    words: list[str],
-    payloads: list[str],
-    *,
-    is_git: bool,
-    heredoc_files: dict[str, str],
-    fail_closed_body_file: bool,
-) -> None:
-    """Extract ``--body-file``/``--file``/``-F`` style file payloads.
-
-    The git-style ``-F <path>`` form is a file reference ONLY for the
-    ``git`` command (codex round-3 #6 — ``gh api -F body=x`` is a field
-    assignment, NOT a file reference). The ``is_git`` flag scopes the
-    short-form ``-F`` reader.
-
-    ``heredoc_files`` maps a path written by a ``> path <<EOF … EOF``
-    redirect earlier in the same command to its body, so a ``-F <path>``
-    reference resolves to the in-command heredoc when the file does not
-    exist on disk yet (#126).
-
-    ``fail_closed_body_file`` decides what an UNREADABLE ``gh``/``glab`` body
-    file does — ``True`` (the destination-aware gates) appends the fail-closed
-    sentinel, ``False`` (the quote scanner) appends nothing (#126). The git
-    ``-F`` commit-message path always fails closed regardless.
-    """
-    i = 0
-    n = len(words)
-    while i < n:
-        word = words[i]
-        if word in _BODY_FILE_FLAG_NAMES and i + 1 < n:
-            _append_file_payload(words[i + 1], payloads, heredoc_files, fail_closed=fail_closed_body_file)
-            i += 2
-            continue
-        attached: str | None = None
-        for flag in _BODY_FILE_FLAG_NAMES:
-            attached = _attached_value(word, flag + "=")
-            if attached is not None:
-                _append_file_payload(attached, payloads, heredoc_files, fail_closed=fail_closed_body_file)
-                break
-        if attached is not None:
-            i += 1
-            continue
-        if is_git and word == "-F" and i + 1 < n:
-            _append_file_payload(words[i + 1], payloads, heredoc_files, fail_closed=True)
-            i += 2
-            continue
-        if is_git:
-            attached = _attached_value(word, "-F")
-            if attached is not None:
-                _append_file_payload(attached, payloads, heredoc_files, fail_closed=True)
-                i += 1
-                continue
-        i += 1
-
-
-def _append_file_payload(path: str, payloads: list[str], heredoc_files: dict[str, str], *, fail_closed: bool) -> None:
-    """Append the body referenced by a ``-F``/``--file``/``--body-file`` path.
-
-    Resolution order: the on-disk file, then an in-command heredoc that
-    writes to that path (``cat > path <<EOF … EOF``), then the
-    ``fail_closed`` branch. The heredoc fallback closes the #126 false
-    positive where a body written to a temp file and committed via ``-F``
-    in the same command was unreadable at PreToolUse scan time (the hook
-    runs BEFORE the file is created).
-
-    ``fail_closed`` selects what an unresolvable path does. ``True`` appends
-    the fail-closed sentinel: the ``git commit -F <path>`` commit-message path
-    always uses it (#1207), as does a ``gh``/``glab`` body file for the
-    destination-aware banned-terms / bare-reference scanners, so a PUBLIC post
-    whose body the gate cannot read hard-blocks rather than slip through unread
-    (a destination-internal post is skipped before the payload is scanned, so
-    the sentinel never over-blocks it). ``False`` appends NOTHING — the quote
-    scanner keeps a drafted-but-absent ``gh``/``glab`` body file as
-    "needs-inline", not a fail-closed HIGH (#126).
-    """
-    content = _read_file_arg(path)
-    if content is None:
-        content = heredoc_files.get(path)
-    if content is not None:
-        payloads.append(content)
-    elif fail_closed:
-        payloads.append(FAIL_CLOSED_SENTINEL)
-
-
-def _handle_api_input(arg: str, payloads: list[str]) -> None:
-    """Read a ``--input`` argument: stdin or missing file → fail closed."""
-    if arg == "-":
-        payloads.append(FAIL_CLOSED_SENTINEL)
-        return
-    content = _read_file_arg(arg)
-    if content is None:
-        payloads.append(FAIL_CLOSED_SENTINEL)
-        return
-    payloads.append(content)
-    payloads.extend(_json_body_fields(content))
-
-
-def _walk_api_fields(words: list[str], payloads: list[str]) -> None:
-    """Extract ``-f``/``-F``/``--field``/``--raw-field`` ``body=`` assignments.
-
-    Also handles ``--input <file>`` / ``--input -`` (stdin → fail closed)
-    and ``--input <missing>`` (fail closed). Field assignments other than
-    ``body=`` are ignored.
-    """
-    field_flags = _API_FIELD_SHORT_FLAGS | _API_FIELD_LONG_FLAGS
-    i = 0
-    n = len(words)
-    while i < n:
-        word = words[i]
-        if word in field_flags and i + 1 < n:
-            _handle_field_assignment(words[i + 1], payloads)
-            i += 2
-            continue
-        if word == "--input" and i + 1 < n:
-            _handle_api_input(words[i + 1], payloads)
-            i += 2
-            continue
-        attached = _attached_value(word, "--input=")
-        if attached is not None:
-            _handle_api_input(attached, payloads)
-        i += 1
-
-
-def _handle_field_assignment(arg: str, payloads: list[str]) -> None:
-    """Parse a ``-F body=value`` style argument and append the value.
-
-    The ``body=`` prefix is required — other field names (``title=``,
-    etc.) are not body-bearing and are ignored.
-    """
-    if "=" not in arg:
-        return
-    name, _, value = arg.partition("=")
-    if name == "body":
-        payloads.append(value)
-
-
 # ── Command-segment walking ─────────────────────────────────────────
 
 
@@ -534,19 +190,19 @@ def _walk_command_segment(
     first, _ = _first_two_words(segment)
     # All segments get the generic body-flag walker since gh, glab, git,
     # and t3 all accept ``--body``/``--message``/``-m``/``-b``.
-    _walk_body_flags(words, payloads)
-    _walk_body_file_flags(
+    walk_body_flags(words, payloads)
+    walk_body_file_flags(
         words,
         payloads,
-        is_git=(first == "git"),
+        leader=first,
         heredoc_files=heredoc_files,
         fail_closed_body_file=fail_closed_body_file,
     )
     # ``gh api`` / ``glab api`` field assignments.
     if first in {"gh", "glab"}:
-        _walk_api_fields(words, payloads)
+        walk_api_fields(words, payloads)
     if first == "curl":
-        _walk_curl_args(words, payloads)
+        walk_curl_args(words, payloads)
 
 
 # ── Body extraction ─────────────────────────────────────────────────
@@ -575,14 +231,14 @@ def extract_bash_payload(command: str, *, fail_closed_body_file: bool = False) -
     cannot read hard-blocks instead of slipping through unread.
     """
     parts: list[str] = []
-    heredoc_files = _heredoc_file_bodies(command)
+    heredoc_files = heredoc_file_bodies(command)
     tokens = tokenize(command)
     for segment in split_commands(tokens):
         _walk_command_segment(segment, parts, heredoc_files, fail_closed_body_file=fail_closed_body_file)
     # Heredocs still need to be parsed against the raw command — the
     # lexer treats them as regular content since heredoc bodies live on
-    # subsequent physical lines. The regex below tolerates that shape.
-    parts.extend(match.group(2) for match in _HEREDOC_RE.finditer(command))
+    # subsequent physical lines. The matcher below tolerates that shape.
+    parts.extend(heredoc_inline_bodies(command))
     # A forge call hidden inside an interpreter / wrapper argument
     # (``sh -c "gh ... --body X"``, ``eval``, ``ssh host gh``, ``xargs gh``)
     # carries its body in an opaque token the walkers cannot descend into; the
@@ -596,28 +252,6 @@ def extract_bash_payload(command: str, *, fail_closed_body_file: bool = False) -
 # ── Secret-scan surfaces (#1672) ────────────────────────────────────
 
 
-def _api_field_values(words: list[str]) -> list[str]:
-    """Return EVERY ``-f``/``-F``/``--field``/``--raw-field`` field VALUE.
-
-    The body extractor keeps only ``body=`` assignments; a secret can equally
-    live in a ``-f title=`` or any other field of a ``gh api`` / ``glab api``
-    call, so the secret scan reads every field value (the part after ``=``)
-    regardless of field name. Bare values (no ``=``) are kept as-is.
-    """
-    field_flags = _API_FIELD_SHORT_FLAGS | _API_FIELD_LONG_FLAGS
-    values: list[str] = []
-    i = 0
-    n = len(words)
-    while i < n:
-        word = words[i]
-        if word in field_flags and i + 1 < n:
-            values.append(words[i + 1].partition("=")[2] or words[i + 1])
-            i += 2
-            continue
-        i += 1
-    return values
-
-
 def extract_secret_scan_text(command: str) -> str:
     """Concatenate EVERY surface a secret must be blocked on, regardless of destination.
 
@@ -626,14 +260,15 @@ def extract_secret_scan_text(command: str) -> str:
     is about. This widens the secret check beyond :func:`extract_bash_payload`
     to also cover the title / commit-subject fragments
     (:func:`extract_title_fragments`) and every ``gh``/``glab api`` field value
-    (:func:`_api_field_values`), so :func:`publish_surface.contains_secret`
-    sees them before the destination skip can short-circuit a scan.
+    (:func:`_arg_body_walkers.api_field_values`), so
+    :func:`publish_surface.contains_secret` sees them before the destination
+    skip can short-circuit a scan.
     """
     parts = [extract_bash_payload(command, fail_closed_body_file=False)]
     parts.extend(extract_title_fragments(command))
     for words in segment_word_lists(command):
         if words[0] in {"gh", "glab"}:
-            parts.extend(_api_field_values(words))
+            parts.extend(api_field_values(words))
     return "\n".join(part for part in parts if part)
